refactor - bottom sheet & map hooks
